requirements.txt
requests
pandas
pytest
python-dotenv